+++ src/ScummRp/scummrp.cpp
@@ -493,6 +493,9 @@ bool ScummRp::_readOption(const char *arg, char *pendingParams)
 			case 'p':
 				ScummRp::_queueParam(pendingParams, c);
 				break;
+			default:
+				ScummIO::fatal(xsprintf("Unrecognized \"-%c\" option", c));
+				break;
 			}
 		}
 	}